--- a/Common/Algorithm/Framework/Alphas/Analysis/InsightManager.cs
+++ b/Common/Algorithm/Framework/Alphas/Analysis/InsightManager.cs
@@ -211,7 +211,9 @@ namespace QuantConnect.Algorithm.Framework.Alphas.Analysis
         /// </summary>
         private void UpdateScores(ReadOnlySecurityValuesCollection securityValuesCollection)
         {
-            var removals = new List<InsightAnalysisContext>();
+            // for performance be lazy to initialize collection
+            List<InsightAnalysisContext> removals = null;
+
             foreach (var context in _openInsightContexts)
             {
                 // was this insight period closed before we update the times?
@@ -255,6 +257,10 @@ namespace QuantConnect.Algorithm.Framework.Alphas.Analysis
                     var id = context.Insight.Id;
                     _closedInsightContexts[id] = context;
 
+                    if (removals == null)
+                    {
+                        removals = new List<InsightAnalysisContext>();
+                    }
                     removals.Add(context);
                 }
 
@@ -262,7 +268,10 @@ namespace QuantConnect.Algorithm.Framework.Alphas.Analysis
                 _updatedInsightContexts.Add(context);
             }
 
-            _openInsightContexts.RemoveWhere(removals.Contains);
+            if (removals != null)
+            {
+                _openInsightContexts.RemoveWhere(removals.Contains);
+            }
         }
 
         /// <summary>Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.</summary>