--- a/Common/Securities/DynamicSecurityData.cs
+++ b/Common/Securities/DynamicSecurityData.cs
@@ -36,16 +36,18 @@ namespace QuantConnect.Securities
         private static readonly MethodInfo GetPropertyMethodInfo = typeof(DynamicSecurityData).GetMethod("GetProperty");
 
         private readonly IRegisteredSecurityDataTypesProvider _registeredTypes;
-        private readonly ConcurrentDictionary<string, Lazy<object>> _storage = new ConcurrentDictionary<string, Lazy<object>>();
         private readonly ConcurrentDictionary<Type, Type> _genericTypes = new ConcurrentDictionary<Type, Type>();
+        private readonly SecurityCache _cache;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DynamicSecurityData"/> class
         /// </summary>
         /// <param name="registeredTypes">Provides all the registered data types for the algorithm</param>
-        public DynamicSecurityData(IRegisteredSecurityDataTypesProvider registeredTypes)
+        /// <param name="cache">The security cache</param>
+        public DynamicSecurityData(IRegisteredSecurityDataTypesProvider registeredTypes, SecurityCache cache)
         {
             _registeredTypes = registeredTypes;
+            _cache = cache;
         }
 
         /// <summary>Returns the <see cref="T:System.Dynamic.DynamicMetaObject" /> responsible for binding operations performed on this object.</summary>
@@ -61,31 +63,23 @@ namespace QuantConnect.Securities
         /// </summary>
         public bool HasData<T>()
         {
-            return _storage.ContainsKey(typeof(T).Name);
+            return _cache.HasData(typeof(T));
         }
 
         /// <summary>
         /// Gets whether or not this dynamic data instance has a property with the specified name.
-        /// This is a case-insensitve search.
+        /// This is a case-insensitive search.
         /// </summary>
         /// <param name="name">The property name to check for</param>
         /// <returns>True if the property exists, false otherwise</returns>
         public bool HasProperty(string name)
         {
-            return _storage.ContainsKey(name);
-        }
-
-        /// <summary>
-        /// Stores the list of data using <paramref name="dataType"/>'s <see cref="Type.Name"/> as the key
-        /// </summary>
-        /// <typeparam name="T">Statically known type of each item in the list. This may just be BaseData
-        /// which is why this method supports explicitly providing the data type to use for the key</typeparam>
-        /// <param name="dataType">The runtime type of each item in the list. The name of this type is used
-        /// as the key for the data</param>
-        /// <param name="data">The data to be stored</param>
-        public void StoreData<T>(Type dataType, IReadOnlyList<T> data)
-        {
-            StoreData(typeof(T), dataType, (IList)data);
+            Type type;
+            if (_registeredTypes.TryGetType(name, out type))
+            {
+                return _cache.HasData(type);
+            }
+            return false;
         }
 
         /// <summary>
@@ -141,10 +135,10 @@ namespace QuantConnect.Securities
         /// <param name="name">The property name to set</param>
         /// <param name="value">The new property value</param>
         /// <returns>Returns the input value back to the caller</returns>
-        public object SetProperty(string name, Lazy<object> value)
+        [Obsolete("DynamicSecurityData is a view of the SecurityCache. It is readonly, properties can not be set")]
+        public object SetProperty(string name, object value)
         {
-            _storage[name] = value;
-            return value;
+            throw new InvalidOperationException("DynamicSecurityData is a view of the SecurityCache. It is readonly, properties can not be set");
         }
 
         /// <summary>
@@ -154,24 +148,22 @@ namespace QuantConnect.Securities
         /// <returns>object value of BaseData</returns>
         public object GetProperty(string name)
         {
-            Lazy<object> value;
-            if (_storage.TryGetValue(name, out value))
-            {
-                return value.Value;
-            }
-
             // check to see if the requested name matches one of the algorithm registered data types and if
             // so, we'll return a new empty list. this precludes us from always needing to check HasData<T>
             Type type;
             if (_registeredTypes.TryGetType(name, out type))
             {
+                IReadOnlyList<BaseData> data;
+                if (_cache.TryGetValue(type, out data))
+                {
+                    return data;
+                }
+
                 var listType = GetGenericListType(type);
                 return Activator.CreateInstance(listType);
             }
 
-            var keys = _storage.Keys.OrderBy(k => k);
-            throw new KeyNotFoundException($"Property with name '{name}' does not exist. Properties: {string.Join(", ", keys)}");
-
+            throw new KeyNotFoundException($"Property with name '{name}' does not exist.");
         }
 
         /// <summary>
@@ -199,7 +191,6 @@ namespace QuantConnect.Securities
                 {
                     list.Add(baseData);
                 }
-                StoreData(type, type, list);
                 return list;
             }
 
@@ -207,38 +198,6 @@ namespace QuantConnect.Securities
                 $"Expected a list with type '{type.GetBetterTypeName()}' " +
                 $"but found type '{data.GetType().GetBetterTypeName()}"
             );
-        }
-
-        /// <summary>
-        /// Store data implementation that covers both Python and C#
-        /// </summary>
-        private void StoreData(Type type, Type dataType, IList data)
-        {
-            Lazy<object> lazyData;
-
-            // this would, for example, be 'Bitcoin' or 'TradeBar'
-            if (type == dataType)
-            {
-                lazyData = new Lazy<object>(() => data);
-            }
-            else
-            {
-                lazyData = new Lazy<object>(() => {
-                    // common case where it's a List<BaseData> but dataType == TradeBar
-                    // create a List<TradeBar> so that when accessed via GetAll<T> or .TradeBar
-                    // the types line up as expected
-                    var listType = GetGenericListType(dataType);
-                    var list = (IList)Activator.CreateInstance(listType);
-                    foreach (var datum in data)
-                    {
-                        // if the element type and dataType aren't in alignment we'll get an invalid cast exception here
-                        list.Add(datum);
-                    }
-
-                    return list;});
-            }
-
-            SetProperty(dataType.Name, lazyData);
         }
 
         private Type GetGenericListType(Type type)