--- a/Common/Securities/RegisteredSecurityDataTypesProvider.cs
+++ b/Common/Securities/RegisteredSecurityDataTypesProvider.cs
@@ -16,7 +16,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace QuantConnect.Securities
 {
@@ -31,7 +30,7 @@ namespace QuantConnect.Securities
         /// </summary>
         public static readonly IRegisteredSecurityDataTypesProvider Null = new RegisteredSecurityDataTypesProvider();
 
-        private readonly Dictionary<string, Type> _types = new Dictionary<string, Type>();
+        private readonly Dictionary<string, Type> _types = new Dictionary<string, Type>(StringComparer.OrdinalIgnoreCase);
 
         /// <summary>
         /// Registers the specified type w/ the provider
@@ -69,19 +68,7 @@ namespace QuantConnect.Securities
         /// </summary>
         public bool TryGetType(string name, out Type type)
         {
-            if (!_types.TryGetValue(name, out type))
-            {
-                // lets try a case insensitive search
-                var kvp = _types.FirstOrDefault(pair => pair.Key.ToLowerInvariant() == name.ToLowerInvariant());
-                if (kvp.Equals(default(KeyValuePair<string, Type>)))
-                {
-                    return false;
-                }
-
-                type = kvp.Value;
-                return true;
-            }
-            return true;
+            return _types.TryGetValue(name, out type);
         }
     }
 }