… Common/Securities/SecurityCacheProvider.cs
… Common/Securities/SecurityService.cs
… Engine/DataFeeds/DataManager.cs
… Engine/DataFeeds/Enumerators/SynchronizingEnumerator.cs
… Engine/HistoricalData/BrokerageHistoryProvider.cs
… Engine/HistoricalData/SubscriptionDataReaderHistoryProvider.cs
… Jupyter/QuantBook.cs
… Tests/Algorithm/AlgorithmRegisterIndicatorTests.cs
… Tests/Algorithm/AlgorithmUniverseSettingsTests.cs
… Tests/Algorithm/Framework/Alphas/OrderBasedInsightGeneratorTests.cs
… Tests/Algorithm/Framework/Portfolio/BlackLittermanOptimizationPortfolioConstructionModelTests.cs
… Tests/Algorithm/Framework/Portfolio/ConfidenceWeightedPortfolioConstructionModelTests.cs
… Tests/Algorithm/Framework/Portfolio/DuplicateKeyPortfolioConstructionModelTests.cs
… Tests/Algorithm/Framework/Portfolio/EqualWeightingPortfolioConstructionModelTests.cs
… Tests/Algorithm/Framework/Portfolio/InsightWeightingPortfolioConstructionModelTests.cs
… Tests/Algorithm/Framework/Risk/MaximumDrawdownPercentPerSecurityTests.cs
… Tests/Brokerages/Bitfinex/BitfinexFeeModelTests.cs
… Tests/Brokerages/BrokerageTests.cs
… Tests/Brokerages/GDAX/GDAXTestHelpers.cs
… Tests/Brokerages/InteractiveBrokers/InteractiveBrokersBrokerageAdditionalTests.cs
… Tests/Brokerages/InteractiveBrokers/InteractiveBrokersBrokerageTests.cs
… Tests/Brokerages/Paper/PaperBrokerageTests.cs
… Tests/Common/Benchmarks/SecurityBenchmarkTests.cs
… Tests/Common/Brokerages/BitfinexBrokerageModelTests.cs
… Tests/Common/Brokerages/FxcmBrokerageModelTests.cs
… Tests/Common/Data/UniverseSelection/SecurityChangesTests.cs
… Tests/Common/Data/UniverseSelection/UniverseTests.cs
… Tests/Common/IsolatorLimitResultProviderTests.cs
… Tests/Common/Orders/Fees/AlphaStreamsFeeModelTests.cs
… Tests/Common/Orders/Fees/InteractiveBrokersFeeModelTests.cs