<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
~PR on top of https://github.com/QuantConnect/Lean/pull/3714 performance improvements v2~
- `DynamicSecurityData` will be a view into the `SecurityCache` instance
- Refactors for `Security` and `SecurityCache` to avoid storing twice
the same data points in the data type cache
- Adding `SecurityCacheProvider` this class allows for two different
`Security` to share the same data type cache through different instance
of `SecurityCache`. This is used to directly access custom data types
through their underlying in a performant manner

Performance tests `BasicTemplateBenchmark C#`:
 ~`pr v2  https://github.com/QuantConnect/Lean/pull/3714`~ `master`
27.89 seconds at 60k data points per second. Processing total of 1,664,679 data points.
27.49 seconds at 61k data points per second. Processing total of 1,664,679 data points.
27.89 seconds at 60k data points per second. Processing total of 1,664,679 data points.
`this pr`
23.88 seconds at 70k data points per second. Processing total of 1,664,679 data points.
23.68 seconds at 70k data points per second. Processing total of 1,664,679 data points.
24.09 seconds at 69k data points per second. Processing total of 1,664,679 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3739
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Need for speed
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->